Setting up infrastructure for kinetic solvers. Ported over most of the files from the earlier MOOSE, still need to complete the implementation of KineticHub. It all compiles but no tests yet.
git-svn-id: https://svn.code.sf.net/p/moose/code/moose/trunk@96 d1858a74-f715-0410-99c9-c66fe9cce3d6

+++ kinetics/Molecule.h
@@ -26,7 +26,7 @@ class Molecule
 		static void setN( const Conn& c, double value );
 		static double getN( const Element* e );
 		static void setMode( const Conn& c, int value );
-		static double getMode( const Element* e );
+		static int getMode( const Element* e );
 		double localGetConc() const;
 		static double getConc( const Element* e );
 		void localSetConc( double value );
